--- Assets/Editor/DragonGodTool/ExportCatalogueFile.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 646c23ee87bf442499540eb8f068c55d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
