+++ Assets/Resources/HotUpdateLoadingPanel.prefab
@@ -65,7 +65,7 @@ MonoBehaviour:
   m_OnCullStateChanged:
     m_PersistentCalls:
       m_Calls: []
-  m_Sprite: {fileID: 21300000, guid: e08b250c1000bf94f8e7c8e9016cae6a, type: 3}
+  m_Sprite: {fileID: 21300000, guid: 3b3f37beb87aa344c978db6273db48e2, type: 3}
   m_Type: 0
   m_PreserveAspect: 1
   m_FillCenter: 1
