--- CatalogueFiles/Android/CatalogueFile.txt
@@ -1,1 +1,0 @@
-{"Config/Client/Player.bin":"CDF7988972A7C33F886D6684A6EC7F8C","Config/Client/RoleRune.bin":"99E24C023BF0D5E34EBAC322271A3FE0","Dll/HotUpdate.dll":"68822D379AA559CA18C3B288F738FF29","Assets/UpdateAssets/Animtion/Play.anim":"1060168B2D10474139526049688954C4","Assets/UpdateAssets/Atlas/Atlas00/Atlas00.spriteatlasv2":"A34CDDEFACC56BC55132BEDF50524F0D","Assets/UpdateAssets/Atlas/Atlas00/Texture/00_alpha.png":"F794D86DB967A5D33D75945EC9DAB526","Assets/UpdateAssets/Atlas/Atlas00/Texture/00_icon.png":"3139855E701DDF18C2D9CB1C76357C7D","Assets/UpdateAssets/Atlas/Atlas00/Texture/00_login.png":"FC8466798550085C886AF43A44C78530","Assets/UpdateAssets/Atlas/Atlas01/Atlas01.spriteatlasv2":"3907E867EBE58BA9833B6C483F99C9F0","Assets/UpdateAssets/Atlas/Atlas01/Texture/01_alpha.png":"F794D86DB967A5D33D75945EC9DAB526","Assets/UpdateAssets/Atlas/Atlas01/Texture/01_AnXing.png":"B8DE2A5968B0F6DAF361D7A87A26E6B0","Assets/UpdateAssets/Atlas/Atlas01/Texture/01_Arrowright.png":"44181B9EFB1693AA78D295236ED2B3B1","Assets/UpdateAssets/Atlas/Atlas01/Texture/01_BeiBao002.png":"0B2350F7332E02CCD358BBEE77A51BCD","Assets/UpdateAssets/Atlas/Atlas01/Texture/01_BiaoTiDian.png":"D6040117F1D5C46924EC8E808C84FFDD","Assets/UpdateAssets/Atlas/Atlas01/Texture/01_BiaoTiFenGeXian.png":"1A6316D987C63E7CD53B5F2521817D66","Assets/UpdateAssets/Atlas/Atlas01/Texture/01_BiaoTiZS030.png":"EAE072365B354444412287E3210C1339","Assets/UpdateAssets/Atlas/Atlas01/Texture/01_CeBianFenGeXian.png":"C35421A48EB1426336FF0CD05400ED3A","Assets/UpdateAssets/Atlas/Atlas01/Texture/01_Cheng1.png":"B4FFD1E024E0AD0F777111812548F664","Assets/UpdateAssets/Atlas/Atlas01/Texture/01_Cheng2.png":"A9534947F0EE69546BD970EE183916DD","Assets/UpdateAssets/Atlas/Atlas01/Texture/01_Cheng3.png":"616E4907F2FBCBDF2B405A46DEFF86F7","Assets/UpdateAssets/Atlas/Atlas01/Texture/01_ChengYuan.png":"890541E3EAD005EAE69B601BBE7C1121","Assets/UpdateAssets/Atlas/Atlas01/Texture/01_dian.png":"F164A657C0B5615E23B431829BDE752E","Assets/UpdateAssets/Atlas/Atlas01/Texture/01_dian2.png":"871184B74DB0026A367FFC7199FAA71C","Assets/UpdateAssets/Atlas/Atlas01/Texture/01_FenYeFenGeXian.png":"BC764EA22431F3C21EB7474EA776EE19","Assets/UpdateAssets/Atlas/Atlas01/Texture/01_HuoBiJia.png":"0AAF0CF6D4DF9453EDFBA8D25FE21333","Assets/UpdateAssets/Atlas/Atlas01/Texture/01_JianTou2.png":"DD59DD849D6355A79E48063E965CACCE","Assets/UpdateAssets/Atlas/Atlas01/Texture/01_JinDuTiao01.png":"E6DA94627CDA42D5E38D65526AA5FD4A","Assets/UpdateAssets/Atlas/Atlas01/Texture/01_JinDuTiao02.png":"CC0C4AB78FF4EA8B9BF36D1226B1073F","Assets/UpdateAssets/Atlas/Atlas01/Texture/01_JinDuTiao03.png":"527FF92CE8D54FDE8F14CDB501094C04","Assets/UpdateAssets/Atlas/Atlas01/Texture/01_Lan1.png":"29D69AEF792F0007C41A8286C6A944CE","Assets/UpdateAssets/Atlas/Atlas01/Texture/01_Lan2.png":"275248D3F0DF871F656FBE8A79872156","Assets/UpdateAssets/Atlas/Atlas01/Texture/01_Lan3.png":"78CADEE6F6C424E0CD1DB5257E5BC3FC","Assets/UpdateAssets/Atlas/Atlas01/Texture/01_QiPao2.png":"6F9AD8B35E3182EB1506453C44D5607B","Assets/UpdateAssets/Atlas/Atlas01/Texture/01_SaoDangZhuangTai1.png":"3A394D3CB608005B2A24CA4B61437A33","Assets/UpdateAssets/Atlas/Atlas01/Texture/01_SaoDangZhuangTai2.png":"D4BC7E25F98049A12C64FB8FEC6621D7","Assets/UpdateAssets/Atlas/Atlas01/Texture/01_TanChuangFenGeXian.png":"058B72BD5BBD13D53774B60F2D6DBF44","Assets/UpdateAssets/Atlas/Atlas01/Texture/01_TanChuangJianTou.png":"A619411261009FCDC8E015DB4DF98D58","Assets/UpdateAssets/Atlas/Atlas01/Texture/01_TanHao.png":"7B899D72DC7DB0BBAB14177DB846325B","Assets/UpdateAssets/Atlas/Atlas01/Texture/01_tuodong.png":"8D538E0B8A08CB81B3791CA908969F42","Assets/UpdateAssets/Atlas/Atlas01/Texture/01_WenHao.png":"1C9388EC4AEC9B9EAE7B62600634857A","Assets/UpdateAssets/Atlas/Atlas02/Atlas02.spriteatlasv2":"D2D5732E045BC3EA0B6F709698CCA91C","Assets/UpdateAssets/Atlas/Atlas02/Texture/02_alpha.png":"F794D86DB967A5D33D75945EC9DAB526","Assets/UpdateAssets/Atlas/Atlas02/Texture/02_FunOpen10.png":"817FDBBA0BBE2EE86BC18FB3F4A03604","Assets/UpdateAssets/Atlas/Atlas02/Texture/02_FunOpen11.png":"262A1D2E1889B9383F9789F385BD3B02","Assets/UpdateAssets/Atlas/Atlas02/Texture/02_FunOpen12.png":"48E4E44DD0D5F23DAC5A6B36695E961C","Assets/UpdateAssets/Atlas/Atlas02/Texture/02_FunOpen13.png":"D79DC337E9A3D724AE37E0C34ADA3219","Assets/UpdateAssets/Atlas/Atlas02/Texture/02_rwtx1.png":"7B0623347C8AC42E1704490BE47169BB","Assets/UpdateAssets/Atlas/Atlas02/Texture/02_rwtx2.png":"7C56AA181CB832A5ED388ABF04BAC321","Assets/UpdateAssets/Atlas/Atlas02/Texture/02_rwtx3.png":"7DCF48FDAF6252AF0C318C895B6EBB2C","Assets/UpdateAssets/Atlas/Atlas02/Texture/02_rwtx4.png":"530ECBB62FA2B43BD90DE5C35A91A3AA","Assets/UpdateAssets/Atlas/Atlas02/Texture/02_rwtx5.png":"51083AC1AD9CD50547E5A3276251A197","Assets/UpdateAssets/Atlas/Atlas02/Texture/02_rwtx6.png":"C7D2B69CCFE610850C9A129B6504AD75","Assets/UpdateAssets/Atlas/Atlas02/Texture/02_rwtx7.png":"4116A33DE877492574CF927A06F750BF","Assets/UpdateAssets/Atlas/Atlas02/Texture/02_rwtx8.png":"CC802B24B680A136590B3B49E7C4BDE8","Assets/UpdateAssets/Atlas/Atlas02/Texture/02_rwtx9.png":"BD5D8897CDBF80DD235C8172A864FC6B","Assets/UpdateAssets/Audios/music.mp3":"A5F60B2C8417BFBAD484DCC86F0895A0","Assets/UpdateAssets/Materials/Grayscale/GrayscaleMaterial.mat":"EBF417B17E22F2F6A6C01FB50BCE1986","Assets/UpdateAssets/Materials/Grayscale/GrayscaleShader.shader":"255EFD0E2605EAD97288ED57747FD823","Assets/UpdateAssets/Materials/UIMaskChildren/UIMaskChildrenMaterial.mat":"A20D633F19C3932131091118DF3BD12C","Assets/UpdateAssets/Materials/UIMaskChildren/UIMaskChildrenShader.shader":"5FC0BCAE7513F6FE723DE9082AFF050E","Assets/UpdateAssets/Materials/UIMaskGrayscale/UIMaskGrayscaleMaterial.mat":"EE20F593BA2E202B4A1182DFA4BEE2AD","Assets/UpdateAssets/Materials/UIMaskGrayscale/UIMaskGrayscaleShader.shader":"96AE65E2B48C572351F102B866F1AA91","Assets/UpdateAssets/Materials/UIMaskParent/UIMaskParentMaterial.mat":"4DC2F77B38E06439B0581AB38D89F73F","Assets/UpdateAssets/Materials/UIMaskParent/UIMaskParentShader.shader":"2C95946F2833F64F614A5378E3C8ED82","Assets/UpdateAssets/Png/bg_hghg_wdtx1.png":"E93D885F48043200F3426DB02C7820EC","Assets/UpdateAssets/Png/bg_hghg_wdtx2.png":"D7B8694CB7EC468F319E82D027C1DCC9","Assets/UpdateAssets/Png/bg_hghg_wdtx3.png":"0A1C16CCE8B65217C2978B6177B1A0A7","Assets/UpdateAssets/Prefabs/UI/CommonPanel/GameLoadingPanel.prefab":"57A470C150640099C566620418ABC033","Assets/UpdateAssets/Prefabs/UI/CommonPanel/GameLoginPanel.prefab":"9777F9B49F912C5D54734E48B4FD6A68"}
